requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.5
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.19.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
wrapt==1.12.1
